add epub helper
